+++ Assets/Game/Scripts/SaveModule.cs
@@ -86,7 +86,7 @@ public class SaveModule : MonoBehaviour
         public int virtualWood;
         public int virtualStone;
         [Header("Character")]
-        public int characterCurrentHealth;
+        public float characterCurrentHealth;
         public int characterCurrentExperience;
         public int characterCurrentLevel;
         public int characterExperienceToNextLevel;
